200 partial original-style asset fetch:
For convenience make singleton for the changeover date.  Use singleton to
bifurcate fetch() into old-style and new-style fetches.  For old-style, pass in
a lat/lon to the search API, gotten from the center point of the chosen tile.

Need to hardlink the asset into all the appropriate tile directories.  Also
need to make sure a series of downloads doesn't break given batch-staging.