239 fix 'requested' incompatibility:
The previous update_or_create would use the whole row as the search key,
but try to set status='complete'.  If the row was already present but
with a different status, this would violate uniqueness constraints.  So,
specify unique-together fields as search key and update the rest of the
fields via `default`.