109 conditionally log instead of print():
Clean up most stray print()s (replace with vprint). Rewrite verbose_out and
vprint to use logging insead of print()s if gips thinks it's operating as a
library (that is if gips_script_setup() hasn't been run). Much of this is
borrowed from dhscratch2 (the last known datahandler branch).

This is covered by unit tests, and system tests pass, and a trial logging
config in gaas showed vprint/verbose_outs going to a log file, so it's
believed that it works.